Refs #20 - Begin converting flux component to wrapper (from parent component).
